Fix sidebar menu being hidden on most screen sizes
The sidebar menu option "Inställningar" and the button labeled "Logga ut" was clipped due to the small column width of the menu in most of the screen widths supported by the website. This commit does a quick fix on the subject by increasing the column width of the menu and decreasing the column width of the content. The content should still be given enough space to be usable.

A better fix would be to remove the max-width property entirely from `sm` and `md` bootstrap styles, or make it a percentage of the screen width, but I could not figure out how to go about that in the limited amount I allocated to fixing this issue.