"Fix" bug where logout and admin button was rendered off screen
For some reason, on phones, when the width of the document increased with the menu toggle the height of the document is increased proportionally. This resulted in the menu items at the bottom of the menu being sent way off screen.

Tested on a normal Linux machine with Firefox and Chrome.