Generate fixture with dummy data
This can be expanded. E.g. create a simple argparse interface to enable
more dynamic generation. Currently it is hard coded to create 200
members and 5 years of productions and association years.

Simple instructions is added to the Readme.

There is probably a way to do it by actually importing and using the
models but I needed a quick solution.

Part of #57