Clarify naming of requirements.txt files
When updating the pip requirements files with pip-compile, it's important to update both the prod and dev requirements, which is now made clearer with a specific "dev" and "prod" naming. Previously, one could think that the general "requirements.txt" would always be used and was the only one which needed to be upgraded, which was not the case.